fix decimal

+++ db/migrate/20161206161852_add_price_to_products.rb
@@ -1,5 +1,5 @@
 class AddPriceToProducts < ActiveRecord::Migration
   def change
-    add_column :products, :price_in_cents, :integer
+    add_column :products, :price_in_cents, :decimal,
   end
 end
